auth-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.10.4
pydantic-core==2.27.2
pydantic-settings==2.1.0
email-validator==2.2.0
PyJWT==2.8.0
fastapi-mail==1.4.1